Re-add shift modifer for constraining to 15 degrees

+++ DHShapeMaker/Extensions.cs
@@ -12,7 +12,7 @@ namespace ShapeMaker
             return ConstrainToInterval((float)amount, interval);
         }
 
-        internal static int ConstrainToInterval(this float amount, int interval)
+        internal static int ConstrainToInterval(this double amount, int interval)
         {
             return (int)(Math.Round(amount / interval) * interval);
         }
